BC-68 added location response header
