+++ src/main/java/com/ssutopia/finacial/loanService/controller/LoanTypeController.java
@@ -1,22 +1,24 @@
 package com.ssutopia.finacial.loanService.controller;
 
-import com.ssutopia.finacial.loanService.entity.Loan;
-import com.ssutopia.finacial.loanService.entity.LoanForm;
-import com.ssutopia.finacial.loanService.dto.LoanTypeDto;
-import com.ssutopia.finacial.loanService.entity.LoanType;
-import com.ssutopia.finacial.loanService.service.LoanService;
-import com.ssutopia.finacial.loanService.service.LoanTypeService;
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+
+import javax.validation.Valid;
+
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.ssutopia.finacial.loanService.dto.LoanTypeDto;
+import com.ssutopia.finacial.loanService.entity.LoanType;
+import com.ssutopia.finacial.loanService.service.LoanTypeService;
+
+import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
@@ -25,9 +27,6 @@ import java.util.List;
 public class LoanTypeController {
 	public static final String MAPPING = EndpointConstants.API_V_0_1_LOANTYPES;
 	private final LoanTypeService loanTypeService;
-
-	@Autowired
-	private LoanService loanService;
 
 	// create loan type
 	@PostMapping(consumes = { MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE })
@@ -47,44 +46,4 @@ public class LoanTypeController {
 		return ResponseEntity.ok(LoanType);
 	}
 
-	// receive loan application form, store in db, & print to console
-	@PostMapping(path = "/form", consumes = { MediaType.APPLICATION_JSON_VALUE }, produces = {
-			MediaType.TEXT_PLAIN_VALUE })
-	public ResponseEntity<String> applyForLoan(@RequestBody LoanForm newLoanForm) {
-		System.out.println("Received a new loan application form:");
-		newLoanForm.printFields();
-
-		delay();
-
-		Loan loan = loanService.createNewLoan(newLoanForm);
-		System.out.println("Created a new loan instance:");
-		loan.printFields();
-
-		return new ResponseEntity<>("", HttpStatus.CREATED);
-	}
-
-	// pretend to think for a few seconds while processing the form
-	private void delay() {
-		try {
-			Thread.sleep(2000);
-		} catch (InterruptedException e) {
-			e.printStackTrace();
-		}
-	}
-
-	/*
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 * 
-	 */
 }
